+++ src/main/java/kitchenpos/application/dto/convertor/MenuConvertor.java
@@ -5,8 +5,8 @@ import java.util.stream.Collectors;
 
 import kitchenpos.application.dto.response.MenuProductResponse;
 import kitchenpos.application.dto.response.MenuResponse;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
 
 public class MenuConvertor {
