+++ src/main/java/kitchenpos/application/dto/convertor/OrderConvertor.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import kitchenpos.application.dto.response.OrderResponse;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 
 public class OrderConvertor {
