+++ src/main/java/kitchenpos/repository/MenuRepository.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Repository;
 
 import kitchenpos.dao.MenuDao;
 import kitchenpos.dao.MenuProductDao;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
 
 @Repository
 public class MenuRepository {
