+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -21,7 +21,7 @@ import kitchenpos.dao.TableGroupDao;
 import kitchenpos.dao.fake.FakeOrderDao;
 import kitchenpos.dao.fake.FakeOrderTableDao;
 import kitchenpos.dao.fake.FakeTableGroupDao;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.TableGroup;
 import kitchenpos.domain.fixture.OrderTableFixture;
