+++ src/test/java/kitchenpos/domain/MenuProductsTest.java
@@ -8,6 +8,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kitchenpos.domain.menu.MenuProduct;
+import kitchenpos.domain.menu.MenuProducts;
+
 @DisplayName("MenuProducts 도메인 테스트")
 class MenuProductsTest {
