+++ src/test/java/kitchenpos/domain/OrderTest.java
@@ -9,6 +9,10 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderLineItem;
+import kitchenpos.domain.order.OrderStatus;
+
 @DisplayName("Order 도메인 테스트")
 class OrderTest {
