+++ src/test/java/kitchenpos/domain/fixture/MenuFixture.java
@@ -3,8 +3,8 @@ package kitchenpos.domain.fixture;
 import java.math.BigDecimal;
 import java.util.List;
 
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
 
 @SuppressWarnings("NonAsciiCharacters")
 public class MenuFixture {
