+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -30,7 +30,7 @@ import kitchenpos.application.dto.request.OrderLineItemRequest;
 import kitchenpos.application.dto.request.OrderRequest;
 import kitchenpos.application.dto.request.OrderChangeRequest;
 import kitchenpos.application.dto.response.OrderResponse;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.domain.order.OrderStatus;
 
 @ApiTest
 @DisplayName("Order API 테스트")
